refactor: use standardized redis uri for dialer arguments
