+++ redis_broker.go
@@ -17,35 +17,10 @@ type RedisCeleryBroker struct {
 	workWG      sync.WaitGroup
 }
 
-// NewRedisPool creates pool of redis connections
-func NewRedisPool(host, pass string) *redis.Pool {
-	return &redis.Pool{
-		MaxIdle:     3,
-		IdleTimeout: 240 * time.Second,
-		Dial: func() (redis.Conn, error) {
-			c, err := redis.Dial("tcp", host)
-			if err != nil {
-				return nil, err
-			}
-			if pass != "" {
-				if _, err = c.Do("AUTH", pass); err != nil {
-					c.Close()
-					return nil, err
-				}
-			}
-			return c, err
-		},
-		TestOnBorrow: func(c redis.Conn, t time.Time) error {
-			_, err := c.Do("PING")
-			return err
-		},
-	}
-}
-
-// NewRedisCeleryBroker creates new RedisCeleryBroker
-func NewRedisCeleryBroker(host, pass string) *RedisCeleryBroker {
+// NewRedisCeleryBroker creates new RedisCeleryBroker based on given uri
+func NewRedisCeleryBroker(uri string) *RedisCeleryBroker {
 	return &RedisCeleryBroker{
-		Pool:      NewRedisPool(host, pass),
+		Pool:      NewRedisPool(uri),
 		queueName: "celery",
 	}
 }
