+++ redis_pool.go
@@ -1,0 +1,27 @@
+package gocelery
+
+import (
+	"time"
+
+	"github.com/gomodule/redigo/redis"
+)
+
+
+// NewRedisPoolURI creates pool of redis connections from given uri
+func NewRedisPoolURI(uri string) *redis.Pool {
+	return &redis.Pool{
+		MaxIdle:     3,
+		IdleTimeout: 240 * time.Second,
+		Dial: func() (redis.Conn, error) {
+			c, err := redis.DialURL(uri)
+			if err != nil {
+				return nil, err
+			}
+			return c, err
+		},
+		TestOnBorrow: func(c redis.Conn, t time.Time) error {
+			_, err := c.Do("PING")
+			return err
+		},
+	}
+}
